VCST-2934: Rename export/import to backup/restore functionality
feat: Rename export/import to backup/restore functionality
