+++ src/VirtoCommerce.Platform.Core/PlatformOptions.cs
@@ -39,7 +39,7 @@ namespace VirtoCommerce.Platform.Core
         // Default path to store export files
         public string DefaultExportFolder { get; set; } = "export";
 
-        public string DefaultExportFileName { get; set; } = "exported_{0:yyyyMMddHHmmss}.zip";
+        public string DefaultExportFileName { get; set; } = "vc_backup_{0:yyyyMMddHHmmss}.zip";
 
         // Local path to running process like WkhtmlToPdf
         public string ProcessesPath { get; set; }
