+++ src/VirtoCommerce.Platform.Web/Controllers/Api/PlatformExportImportController.cs
@@ -8,7 +8,6 @@ using Hangfire;
 using Hangfire.Server;
 using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
-using Microsoft.AspNetCore.StaticFiles;
 using Microsoft.Extensions.Options;
 using VirtoCommerce.Platform.Core;
 using VirtoCommerce.Platform.Core.Common;
@@ -165,68 +164,6 @@ namespace VirtoCommerce.Platform.Web.Controllers.Api
             return Ok(retVal);
         }
 
-        [HttpPost]
-        [Route("export")]
-        [Authorize(Permissions.PlatformExport)]
-        public ActionResult<PlatformExportPushNotification> ProcessExport([FromBody] PlatformImportExportRequest exportRequest)
-        {
-            var notification = new PlatformExportPushNotification(_userNameResolver.GetCurrentUserName())
-            {
-                Title = "Platform export task",
-                Description = "starting export...."
-            };
-            _pushNotifier.Send(notification);
-
-            var jobId = BackgroundJob.Enqueue(() => PlatformExportBackgroundAsync(exportRequest, notification, JobCancellationToken.Null, null));
-            notification.JobId = jobId;
-            return Ok(notification);
-        }
-
-        [HttpPost]
-        [Route("import")]
-        [Authorize(Permissions.PlatformImport)]
-        public ActionResult<PlatformImportPushNotification> ProcessImport([FromBody] PlatformImportExportRequest importRequest)
-        {
-            var notification = new PlatformImportPushNotification(_userNameResolver.GetCurrentUserName())
-            {
-                Title = "Platform import task",
-                Description = "starting import...."
-            };
-            _pushNotifier.Send(notification);
-
-            var jobId = BackgroundJob.Enqueue(() => PlatformImportBackgroundAsync(importRequest, notification, JobCancellationToken.Null, null));
-            notification.JobId = jobId;
-
-            return Ok(notification);
-        }
-
-        [HttpPost]
-        [Route("exortimport/tasks/{jobId}/cancel")]
-        public ActionResult Cancel([FromRoute] string jobId)
-        {
-            BackgroundJob.Delete(jobId);
-            return Ok();
-        }
-
-        [HttpGet]
-        [Route("export/download/{fileName}")]
-        [Authorize(Permissions.PlatformExport)]
-        public ActionResult DownloadExportFile([FromRoute] string fileName)
-        {
-            var localPath = GetSafeFullPath(_platformOptions.DefaultExportFolder, fileName);
-
-            //Load source data only from local file system
-            using (System.IO.File.Open(localPath, FileMode.Open))
-            {
-                var provider = new FileExtensionContentTypeProvider();
-                if (!provider.TryGetContentType(localPath, out var contentType))
-                {
-                    contentType = "application/octet-stream";
-                }
-                return PhysicalFile(localPath, contentType);
-            }
-        }
-
         private async Task<IList<SampleDataInfo>> InnerDiscoverSampleDataAsync()
         {
             var sampleDataUrl = _platformOptions.SampleDataUrl;
@@ -342,100 +279,12 @@ namespace VirtoCommerce.Platform.Web.Controllers.Api
             finally
             {
                 await _settingsManager.SetValueAsync(PlatformConstants.Settings.Setup.SampleDataState.Name, SampleDataState.Completed);
-                pushNotification.Description = "Import finished";
+                pushNotification.Description = "Sample data import process completed successfully.";
                 pushNotification.Finished = DateTime.UtcNow;
                 await _pushNotifier.SendAsync(pushNotification);
             }
         }
 
-        public async Task PlatformImportBackgroundAsync(PlatformImportExportRequest importRequest, PlatformImportPushNotification pushNotification, IJobCancellationToken cancellationToken, PerformContext context)
-        {
-            void progressCallback(ExportImportProgressInfo x)
-            {
-                pushNotification.Path(x);
-                pushNotification.JobId = context.BackgroundJob.Id;
-                _pushNotifier.Send(pushNotification);
-            }
-
-            var now = DateTime.UtcNow;
-            try
-            {
-                var cancellationTokenWrapper = new JobCancellationTokenWrapper(cancellationToken);
-
-                var localPath = GetSafeFullPath(_platformOptions.LocalUploadFolderPath, importRequest.FileUrl);
-
-                //Load source data only from local file system
-                using (var stream = new FileStream(localPath, FileMode.Open))
-                {
-                    var manifest = importRequest.ToManifest();
-                    manifest.Created = now;
-                    await _platformExportManager.ImportAsync(stream, manifest, progressCallback, cancellationTokenWrapper);
-                }
-            }
-            catch (JobAbortedException)
-            {
-                //do nothing
-            }
-            catch (Exception ex)
-            {
-                pushNotification.Errors.Add(ex.ExpandExceptionMessage());
-            }
-            finally
-            {
-                pushNotification.Description = "Import finished";
-                pushNotification.Finished = DateTime.UtcNow;
-                await _pushNotifier.SendAsync(pushNotification);
-            }
-        }
-
-        public async Task PlatformExportBackgroundAsync(PlatformImportExportRequest exportRequest, PlatformExportPushNotification pushNotification, IJobCancellationToken cancellationToken, PerformContext context)
-        {
-            void progressCallback(ExportImportProgressInfo x)
-            {
-                pushNotification.Path(x);
-                pushNotification.JobId = context.BackgroundJob.Id;
-                _pushNotifier.Send(pushNotification);
-            }
-
-            try
-            {
-                var fileName = string.Format(_platformOptions.DefaultExportFileName, DateTime.UtcNow);
-                var localTmpFolder = Path.GetFullPath(Path.Combine(_platformOptions.DefaultExportFolder));
-                var localTmpPath = Path.Combine(localTmpFolder, Path.GetFileName(fileName));
-
-                if (!Directory.Exists(localTmpFolder))
-                {
-                    Directory.CreateDirectory(localTmpFolder);
-                }
-
-                if (System.IO.File.Exists(localTmpPath))
-                {
-                    System.IO.File.Delete(localTmpPath);
-                }
-
-                //Import first to local tmp folder because Azure blob storage doesn't support some special file access mode
-                using (var stream = System.IO.File.OpenWrite(localTmpPath))
-                {
-                    var manifest = exportRequest.ToManifest();
-                    await _platformExportManager.ExportAsync(stream, manifest, progressCallback, new JobCancellationTokenWrapper(cancellationToken));
-                    pushNotification.DownloadUrl = $"api/platform/export/download/{fileName}";
-                }
-            }
-            catch (JobAbortedException)
-            {
-                //do nothing
-            }
-            catch (Exception ex)
-            {
-                pushNotification.Errors.Add(ex.ExpandExceptionMessage());
-            }
-            finally
-            {
-                pushNotification.Description = "Export finished";
-                pushNotification.Finished = DateTime.UtcNow;
-                await _pushNotifier.SendAsync(pushNotification);
-            }
-        }
 
         private static string GetSafeFullPath(string basePath, string relativePath)
         {
