+++ src/VirtoCommerce.Platform.Web/wwwroot/Localizations/ru.VirtoCommerce.Platform.json
@@ -10,7 +10,7 @@
       "dynamic-properties": "Динамические свойства",
       "security": "Безопасность",
       "modules": "Модули",
-      "export-import": "Экспорт и импорт",
+      "export-import": "Резервное копирование и Восстановление",
       "user-profile": "Профиль пользователя",
       "assets": "Ресурсы",
       "toggle-favorites": "Shift + Пробел для переключения избранного",
@@ -152,10 +152,10 @@
         }
       },
       "export-main": {
-        "title": "Экспорт данных",
+        "title": "Резервное копирование данных",
         "labels": {
           "platform-entries": "Записи платформы",
-          "choose-modules": "Выберите модули для экспорта",
+          "choose-modules": "Выберите модули для резервного копирования",
           "download-url": "URL для скачивания",
           "errors": "Ошибки",
           "start": "Начало",
@@ -181,21 +181,21 @@
         }
       },
       "import-main": {
-        "title": "Импорт данных",
+        "title": "Восстановление данных",
         "labels": {
-          "start-import": "Начать импорт",
-          "select-file": "Выберите .zip файл для импорта. Файл должен быть создан экспортом данных Commerce Manager",
-          "drag-note": "Перетащите .zip файл для импорта. Файл должен быть создан экспортом данных Commerce Manager",
+          "start-import": "Начать восстановление",
+          "select-file": "Выберите .zip-файл для восстановления. Файл должен быть создан резервным копированием данных Virto Commerce",
+          "drag-note": "Перетащите .zip-файл сюда для восстановления. Файл должен быть создан резервным копированием данных Virto Commerce",
           "upload-failed": "Сбой загрузки",
           "upload-progress": "Прогресс загрузки",
-          "import-progress": "Прогресс импорта",
-          "import-finished": "Импорт завершен",
-          "import-data-information": "Информация об импортируемых данных",
+          "import-progress": "Прогресс восстановления",
+          "import-finished": "Восстановление завершено",
+          "import-data-information": "Информация о восстановленых данных",
           "author": "Автор",
           "data-file-created": "Файл данных создан",
           "created-platform-version": "Создано в версии платформы",
           "platform-entries": "Записи платформы",
-          "choose-modules": "Выберите модули для импорта",
+          "choose-modules": "Выберите модули для восстановления",
           "no-data": "Нет данных",
           "errors": "Ошибки",
           "start": "Начало",
@@ -221,15 +221,15 @@
         }
       },
       "exportImport-main": {
-        "title": "Экспорт и импорт данных",
+        "title": "Резервное копирование и восстановление данных",
         "menu": {
           "export": {
-            "title": "Экспорт",
-            "description": "Данные будут сохранены в файл, чтобы их можно было импортировать позже"
+            "title": "Резервное копирование",
+            "description": "Данные будут сохранены в файле для последующего восстановления"
           },
           "import": {
-            "title": "Импорт",
-            "description": "Импорт данных из экспортированного файла"
+            "title": "Восстановление",
+            "description": "Восстановите данные из файла резервной копии для будущего использования"
           }
         }
       },
@@ -919,9 +919,9 @@
     "platform:security:delete": "Удаление пользователей и ролей",
     "platform:security:verifyEmail": "Изменение состояния подтверждения и отправка письма для подтверждения",
     "background_jobs:manage": "Доступ к консоли Hangfire",
-    "platform:exportImport:access": "Просмотр Экспорта/Импорта в главном меню",
-    "platform:import": "Импорт данных платформы и модулей",
-    "platform:export": "Экспорт данных платформы и модулей",
+    "platform:exportImport:access": "Просмотр Резервного копирования/Восстановления в главном меню",
+    "platform:import": "Восстановить данные платформы и модулей",
+    "platform:export": "Создать резервную копию данных платформы и модулей",
     "platform:security:loginOnBehalf": "Вход от имени клиента"
   },
   "settings": {
